GEF maintance part 1 of many
Removed German East Asia and Indies. The indies will later be changed into something else, as the british will get it. Fixed Adelsverein decisions, though a bit op right now. Fixed the annexing of the Lowlands...hopefully. Added first debuff for Germany. Removed the option to annex Wallonia when going to war with Belgium. Added/modified renaming decisions for the Lowlands. Removed the possiblity for Germany to annex Denmark and Switzerland.