Potential Fix for White Peace
It appears that white peacing is counted as losing the war. Thus, a slight rewrite to (hopefully) assure that this does not trigger the dismantelment of the German Confederation. Also added some time limits on the events, as to prevent future wars from triggering it. ~10 years is given for taking Austria and for Russia to dismantel the German Confederation.