Hawaii Province Update
Update changes the names of the hawaiian provinces to be native hawaiian names, adds positions, makes the main islands a total of four, updates pop data to be as accurate as possible. All the provinces are connected by straits, so they can be traveled across without boats when troops are landed. Exact pop information regarding farmers and artisans and craftsmen I am not sure about but I based it on the base pop ratios we had already.